final 1.0 version

+++ app/readfile.py
@@ -22,7 +22,7 @@ class FileHandler(BaseHandler):
     def post(self):
         user_folder=os.path.join(Settings.UPLOADS,self.current_user.replace('\"','')) + '/'
         os.system('rm -f '+user_folder+'*.*')
-        os.system('rm -rf '+user_folder+'results/')
+        #os.system('rm -rf '+user_folder+'results/')
         check = self.get_argument("check")
         print '+-+-+-+-',check
         id1 = self.get_argument("id")
@@ -31,6 +31,9 @@ class FileHandler(BaseHandler):
         print '+-+-+-+-',xs
         ys = self.get_argument("ysize")
         print '+-+-+-+-',ys
+        listonly = self.get_argument("listonly")
+        print '+-+-+-+-',listonly
+        print 
         print self.current_user
         if check == "file":
             fileinfo = self.request.files["csvfile"][0]
@@ -60,23 +63,37 @@ class FileHandler(BaseHandler):
         os.system(comm)
         mypath = '/static/uploads/'+self.current_user.replace('\"','')+'/results/'
 
-        tiffiles=glob.glob(user_folder+'results/*.tif')
-        titles=[]
-        pngfiles=[]
-        Ntiles = len(tiffiles)
-        for f in tiffiles:
-            title=f.split('/')[-1][:-4]
-            os.system("convert %s %s.png" % (f,f))
-            titles.append(title)
-            pngfiles.append(mypath+title+'.tif.png')
-       
-        os.chdir(user_folder)
-        os.system("tar -zcf results/all.tar.gz results/") 
-        os.chdir(os.path.dirname(__file__))
-        if os.path.exists(user_folder+"list.json"): os.remove(user_folder+"list.json")
-        with open(user_folder+"list.json","w") as outfile:
-            json.dump([dict(name=pngfiles[i],title=titles[i], size=Ntiles) for i in range(len(pngfiles))], outfile, indent=4)
+        allfiles = glob.glob(user_folder+'results/*.*')
+        Fall = open(user_folder+'results/list_all.txt','w')
+        prefix='http://desdev2.cosmology.illinois.edu:8888/static'
+        for ff in allfiles: Fall.write(prefix+ff.split('static')[-1]+'\n')
+        Fall.close()
 
-        self.set_status(200)
-        self.flush()
+        if listonly == 'yes':
+            if os.path.exists(user_folder+"list.json"): os.remove(user_folder+"list.json")
+            with open(user_folder+"list.json","w") as outfile:
+                json.dump('', outfile, indent=4)
+            self.set_status(200)
+            self.flush()
+        else:
+
+            tiffiles=glob.glob(user_folder+'results/*.tif')
+            titles=[]
+            pngfiles=[]
+            Ntiles = len(tiffiles)
+            for f in tiffiles:
+                title=f.split('/')[-1][:-4]
+                os.system("convert %s %s.png" % (f,f))
+                titles.append(title)
+                pngfiles.append(mypath+title+'.tif.png')
+       
+            os.chdir(user_folder)
+            os.system("tar -zcf results/all.tar.gz results/") 
+            os.chdir(os.path.dirname(__file__))
+            if os.path.exists(user_folder+"list.json"): os.remove(user_folder+"list.json")
+            with open(user_folder+"list.json","w") as outfile:
+                json.dump([dict(name=pngfiles[i],title=titles[i], size=Ntiles) for i in range(len(pngfiles))], outfile, indent=4)
+
+            self.set_status(200)
+            self.flush()
         self.finish()
